[dagit] Copyable links to individual instigation ticks
