+++ js_modules/dagit/packages/core/src/graphql/schema.graphql
@@ -1028,6 +1028,7 @@ type InstigationState {
   typeSpecificData: InstigationTypeSpecificData
   runs(limit: Int): [Run!]!
   runsCount: Int!
+  tick(timestamp: Float): InstigationTick
   ticks(dayRange: Int, dayOffset: Int, limit: Int): [InstigationTick!]!
   nextTick: FutureInstigationTick
   runningCount: Int!
